+++ fourtoon-cookie/src/components/global/GlobalSelectionCharacter/GlobalSelectionCharacterStateContext.tsx
@@ -1,14 +1,14 @@
 import { createContext } from 'react';
-import type { Character } from '../../types/character';
+import type { Character } from '../../../types/character';
 
 export interface GlobalSelectionCharacterStateContextProps {
-  selectedCharacter: Character | null;
-  setSelectedCharacter: (character: Character | null) => void;
+	selectedCharacter: Character | null;
+	setSelectedCharacter: (character: Character | null) => void;
 }
 
 export const defaultValueOfGlobalSelectionCharacterStateContextProps: GlobalSelectionCharacterStateContextProps = {
-  selectedCharacter: null,
-  setSelectedCharacter: () => { },
+	selectedCharacter: null,
+	setSelectedCharacter: () => { },
 };
 
 const GlobalSelectionCharacterStateContext = createContext<GlobalSelectionCharacterStateContextProps>(defaultValueOfGlobalSelectionCharacterStateContextProps);
