+++ fourtoon-cookie/src/pages/DiaryWritePage/Header/Header.tsx
@@ -6,7 +6,7 @@ import BackButton from "../../../components/common/BackButton/BackButton";
 import * as S from "./Header.styled";
 import CharacterItem from "../../../components/character/CharacterItem/CharacterItem";
 import { useContext, useEffect } from "react";
-import GlobalSelectionCharacterStateContext from "../../../components/global/GlobalSelectionCharacterStateContext";
+import GlobalSelectionCharacterStateContext from "../../../components/global/GlobalSelectionCharacter/GlobalSelectionCharacterStateContext";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
 import { RootStackParamList } from "../../../constants/routing";
